requirements.txt
anthropic>=0.25.0
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
flask>=3.0.0
higgsfield-client>=0.1.0
Pillow>=10.0.0
